+++ src/backend/application.py
@@ -5,30 +5,41 @@ import io
 import base64
 from PIL import Image
 import numpy as np
-from wtforms import Form, StringField, TextAreaField, PasswordField, validators
-from passlib.hash import sha256_crypt
 import DatabaseConnector
 
 application = Flask(__name__)
 application .secret_key = 'admin'
 socketio = SocketIO(application)
 
-application.config['MYSQL_HOST']='heredb.citwg2mji1tb.us-east-2.rds.amazonaws.com'
-application.config['MYSQL_USER']='admin'
-application.config['MYSQL_PASSWORD']='hereadmin'
-application.config['MYSQL_DB']='here'
-application.config['MYSQL_CURSORCLASS']='DictCursor'
+# Config DB
+db = DatabaseConnector.DatabaseConnector(application, 'heredb.citwg2mji1tb.us-east-2.rds.amazonaws.com',
+                        'admin', 'hereadmin', 'here')
 
-mysql = MySQL(application)
-
+"""
+@application.route('/', methods=['POST', 'GET'])
+def index():
+    return render_template('index.html')
+"""
 @application.route('/')
 def users():
-    cursor = mysql.connection.cursor()
-    try:
-      cursor.execute('SELECT * FROM example')
-      return str(cursor.fetchall())
-    except Exception as e:
-      print("Problem reading query")
+    print("hello")
+    output = db.read_query('SELECT * FROM example')
+    return str(output)
+"""
+@socketio.on('image')
+def image(data):
+    sbuf = io.StringIO()
+    sbuf.write(data)
 
+    # decode and convert into image
+    b = io.BytesIO(base64.b64decode(data))
+    pimg = Image.open(b)
+
+    # converting RGB to BGR, as opencv standards
+    frame = cv2.cvtColor(np.array(pimg), cv2.COLOR_RGB2BGR)
+
+    # Processing here
+    
+"""
 if __name__ == '__main__':
     application.run(debug=True)
